Shuffle files: rename lib/ to pkg/ as first step towards better project layout
