+++ pkg/scionutil/connection.go
