+++ pkg/scionutil/hosts.go
