+++ pkg/scionutil/hosts_test.go
