+++ pkg/scionutil/hosts_test_file
