+++ pkg/scionutil/path_selection.go
