+++ pkg/shttp/README.md
