+++ pkg/shttp/examples/image_server/README.md
