+++ pkg/shttp/examples/image_server/dog.jpg
