+++ pkg/shttp/examples/minimal/README.md
