+++ pkg/shttp/examples/minimal/sample.html
