+++ ssh/scionutils/scionutils.go
@@ -9,7 +9,7 @@ import (
 	"github.com/scionproto/scion/go/lib/snet"
 	"github.com/scionproto/scion/go/lib/snet/squic"
 
-	"github.com/netsec-ethz/scion-apps/lib/scionutil"
+	"github.com/netsec-ethz/scion-apps/pkg/scionutil"
 	"github.com/netsec-ethz/scion-apps/ssh/quicconn"
 )
